Added logic to FileSystem class.
Added methods for retrieving the selected options as either a
List<String>, or a String[].

Added some logic to the FileSystem#listFiles(String, boolean,
ListingOptions) method.
TODO: Add the rest of the logic to the method.

Added FileListingException.
Added new exception that will be thrown if an error is detected while
listing the files in a directory on the device.